# Optimize cohorts for blockwise distribution of groups